结构变动

+++ src/main/java/com/wpc/admin/service/AuthMenuService.java
@@ -3,13 +3,13 @@ package com.wpc.admin.service;
 import java.util.List;
 
 import com.wpc.admin.entity.AuthMenu;
-import com.wpc.common.BaseService;
+import com.wpc.common.base.service.BaseService;
 
 /**
  * 操作相关
  * author wpc
  */
-public interface AuthMenuService extends BaseService<AuthMenu, Integer> {
+public interface AuthMenuService extends BaseService<AuthMenu, Long> {
 	
 	public final static String BEAN_ID="menuService";
